♻️ Replace defaultCode with componentData
