--- doc/src/components/common/DefaultCode/index.js
@@ -1,25 +1,0 @@
-export { defaultCode as CardDefaultCode } from './Card';
-export { defaultCode as ButtonDefaultCode } from './Button';
-export { defaultCode as CheckboxDefaultCode } from './Checkbox';
-export { defaultCode as CheckboxGroupDefaultCode } from './CheckboxGroup';
-export { defaultCode as ToggleDefaultCode } from './Toggle';
-export { defaultCode as RadioButtonGroupDefaultCode } from './RadioButtonGroup';
-export { defaultCode as TextInputDefaultCode } from './TextInput';
-export { defaultCode as SnackbarDefaultCode } from './Snackbar';
-export { defaultCode as AvatarDefaultCode } from './Avatar';
-export { defaultCode as TooltipDefaultCode } from './Tooltip';
-export { defaultCode as AutoCompleteDefaultCode } from './AutoComplete';
-export { defaultCode as BadgeDefaultCode } from './Badge';
-export { defaultCode as DrawerDefaultCode } from './Drawer';
-export { defaultCode as ChipDefaultCode } from './Chip';
-export { defaultCode as MenuDefaultCode } from './Menu';
-export { defaultCode as ModalDefaultCode } from './Modal';
-export { defaultCode as MultiSelectDefaultCode } from './MultiSelect';
-export { defaultCode as NavbarDefaultCode } from './Navbar';
-export { defaultCode as PopoverDefaultCode } from './Popover';
-export { defaultCode as PreloaderDefaultCode } from './Preloader';
-export { defaultCode as SelectDefaultCode } from './Select';
-export { defaultCode as SliderDefaultCode } from './Slider';
-export { defaultCode as CarouselDefaultCode } from './Carousel';
-export { defaultCode as ProgressBarDefaultCode } from './ProgressBar';
-export { defaultCode as TableDefaultCode } from './Table';
