--- doc/src/components/common/DefaultCode/theme.scss
@@ -1,4 +1,0 @@
-@import './Card/style.scss';
-@import './Button/style.scss';
-@import './Avatar/style.scss';
-@import './Navbar/style.scss';
