+++ doc/src/components/common/componentData/Table/theme.scss
