+++ doc/src/components/common/componentData/index.js
@@ -1,0 +1,25 @@
+export { componentData as CardComponentData } from './Card';
+export { componentData as ButtonComponentData } from './Button';
+export { componentData as CheckboxComponentData } from './Checkbox';
+export { componentData as CheckboxGroupComponentData } from './CheckboxGroup';
+export { componentData as ToggleComponentData } from './Toggle';
+export { componentData as RadioButtonGroupComponentData } from './RadioButtonGroup';
+export { componentData as TextInputComponentData } from './TextInput';
+export { componentData as SnackbarComponentData } from './Snackbar';
+export { componentData as AvatarComponentData } from './Avatar';
+export { componentData as TooltipComponentData } from './Tooltip';
+export { componentData as AutoCompleteComponentData } from './AutoComplete';
+export { componentData as BadgeComponentData } from './Badge';
+export { componentData as DrawerComponentData } from './Drawer';
+export { componentData as ChipComponentData } from './Chip';
+export { componentData as MenuComponentData } from './Menu';
+export { componentData as ModalComponentData } from './Modal';
+export { componentData as MultiSelectComponentData } from './MultiSelect';
+export { componentData as NavbarComponentData } from './Navbar';
+export { componentData as PopoverComponentData } from './Popover';
+export { componentData as PreloaderComponentData } from './Preloader';
+export { componentData as SelectComponentData } from './Select';
+export { componentData as SliderComponentData } from './Slider';
+export { componentData as CarouselComponentData } from './Carousel';
+export { componentData as ProgressBarComponentData } from './ProgressBar';
+export { componentData as TableComponentData } from './Table';
