+++ doc/src/components/common/componentList.js
@@ -25,157 +25,157 @@ import ProgressBar from '../../../../src/progressbar';
 import Table from '../../../../src/table';
 
 import {
-  CardDefaultCode,
-  ButtonDefaultCode,
-  CheckboxDefaultCode,
-  CheckboxGroupDefaultCode,
-  ToggleDefaultCode,
-  RadioButtonGroupDefaultCode,
-  TextInputDefaultCode,
-  SnackbarDefaultCode,
-  AvatarDefaultCode,
-  TooltipDefaultCode,
-  AutoCompleteDefaultCode,
-  BadgeDefaultCode,
-  DrawerDefaultCode,
-  ChipDefaultCode,
-  MenuDefaultCode,
-  ModalDefaultCode,
-  MultiSelectDefaultCode,
-  NavbarDefaultCode,
-  PopoverDefaultCode,
-  PreloaderDefaultCode,
-  SelectDefaultCode,
-  SliderDefaultCode,
-  CarouselDefaultCode,
-  ProgressBarDefaultCode,
-  TableDefaultCode,
-} from './DefaultCode';
+  CardComponentData,
+  ButtonComponentData,
+  CheckboxComponentData,
+  CheckboxGroupComponentData,
+  ToggleComponentData,
+  RadioButtonGroupComponentData,
+  TextInputComponentData,
+  SnackbarComponentData,
+  AvatarComponentData,
+  TooltipComponentData,
+  AutoCompleteComponentData,
+  BadgeComponentData,
+  DrawerComponentData,
+  ChipComponentData,
+  MenuComponentData,
+  ModalComponentData,
+  MultiSelectComponentData,
+  NavbarComponentData,
+  PopoverComponentData,
+  PreloaderComponentData,
+  SelectComponentData,
+  SliderComponentData,
+  CarouselComponentData,
+  ProgressBarComponentData,
+  TableComponentData,
+} from './componentData';
 
 export const componentList = [
   {
     name: 'AutoComplete',
     component: AutoComplete,
-    defaultCode: AutoCompleteDefaultCode
+    componentData: AutoCompleteComponentData,
   },
   {
     name: 'Avatar',
     component: Avatar,
-    defaultCode: AvatarDefaultCode,
+    componentData: AvatarComponentData,
   },
   {
     name: 'Badge',
     component: Badge,
-    defaultCode: BadgeDefaultCode,
+    componentData: BadgeComponentData,
   },
   {
     name: 'Button',
     component: Button,
-    defaultCode: ButtonDefaultCode,
+    componentData: ButtonComponentData,
   },
   {
     name: 'Card',
     component: Card,
-    defaultCode: CardDefaultCode,
+    componentData: CardComponentData,
   },
   {
     name: 'Carousel',
     component: Carousel,
-    defaultCode: CarouselDefaultCode,
+    componentData: CarouselComponentData,
   },
   {
     name: 'Checkbox',
     component: Checkbox,
-    defaultCode: CheckboxDefaultCode,
+    componentData: CheckboxComponentData,
   },
   {
     name: 'CheckboxGroup',
     component: CheckboxGroup,
-    defaultCode: CheckboxGroupDefaultCode,
+    componentData: CheckboxGroupComponentData,
   },
   {
     name: 'Chip',
     component: Chip,
-    defaultCode: ChipDefaultCode,
+    componentData: ChipComponentData,
   },
   {
     name: 'Drawer',
     component: Drawer,
-    defaultCode: DrawerDefaultCode,
+    componentData: DrawerComponentData,
   },
   {
     name: 'Menu',
     component: Menu,
-    defaultCode: MenuDefaultCode,
+    componentData: MenuComponentData,
   },
   {
     name: 'Modal',
     component: Modal,
-    defaultCode: ModalDefaultCode,
+    componentData: ModalComponentData,
   },
   {
     name: 'MultiSelect',
     component: MultiSelect,
-    defaultCode: MultiSelectDefaultCode,
+    componentData: MultiSelectComponentData,
   },
   {
     name: 'Navbar',
     component: Navbar,
-    defaultCode: NavbarDefaultCode,
+    componentData: NavbarComponentData,
   },
   {
     name: 'Popover',
     component: Popover,
-    defaultCode: PopoverDefaultCode,
+    componentData: PopoverComponentData,
   },
   {
     name: 'Preloader',
     component: Preloader,
-    defaultCode: PreloaderDefaultCode,
+    componentData: PreloaderComponentData,
   },
   {
     name: 'ProgressBar',
     component: ProgressBar,
-    defaultCode: ProgressBarDefaultCode,
+    componentData: ProgressBarComponentData,
   },
   {
     name: 'RadioButtonGroup',
     component: RadioButtonGroup,
-    defaultCode: RadioButtonGroupDefaultCode,
+    componentData: RadioButtonGroupComponentData,
   },
   {
     name: 'Select',
     component: Select,
-    defaultCode: SelectDefaultCode,
+    componentData: SelectComponentData,
   },
   {
     name: 'Slider',
     component: Slider,
-    defaultCode: SliderDefaultCode,
+    componentData: SliderComponentData,
   },
   {
     name: 'Snackbar',
     component: Snackbar,
-    defaultCode: SnackbarDefaultCode,
+    componentData: SnackbarComponentData,
   },
   {
     name: 'Table',
     component: Table,
-    defaultCode: TableDefaultCode,
+    componentData: TableComponentData,
   },
   {
     name: 'TextInput',
     component: TextInput,
-    defaultCode: TextInputDefaultCode,
+    componentData: TextInputComponentData,
   },
   {
     name: 'Toggle',
     component: Toggle,
-    defaultCode: ToggleDefaultCode,
+    componentData: ToggleComponentData,
   },
   {
     name: 'Tooltip',
     component: Tooltip,
-    defaultCode: TooltipDefaultCode,
+    componentData: TooltipComponentData,
   }
 ];
